requirements.txt
asgiref==3.2.10
autopep8==1.6.0
azure-core==1.24.0
azure-storage-blob==12.12.0
bleach==5.0.0
boto3==1.22.11
botocore==1.25.11
cachetools==5.1.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.50.0
django-cors-headers==3.10.0
django-countries==7.2.1
django-debug-toolbar==3.2.2
django-google-api==1.0.1
django-heroku==0.3.1
django-shopping-cart==0.1
django-storages==1.12.3
django-summernote==0.8.20.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
isodate==0.6.1
jmespath==1.0.0
msrest==0.6.21
numpy==1.21.4
oauthlib==3.2.0
opencv-python==4.5.4.60
Pillow==8.4.0
protobuf==3.20.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2020.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
Serializer==0.2.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
